+++ pages/course/index.js
@@ -1,6 +1,6 @@
-import { Container } from '../../components/layout';
-import Table from '../../components/course/table';
 import Banner from '../../components/course/banner';
+import Table from '../../components/course/table';
+import { Container } from '../../components/layout';
 
 export default function Course() {
   const metaTags = {
@@ -14,21 +14,21 @@ export default function Course() {
   return (
     <Container metaTags={metaTags}>
       <div className="mx-2">
-        <div className="flex justify-center">
+        <div className="flex flex-col items-center">
           <h1 className="text-2xl font-bold capitalize text-gray-900 dark:text-gray-200 md:text-3xl 2xl:text-4xl">
             Solana Development Course
           </h1>
+
+          <Banner />
         </div>
 
         <div className="mx-auto mt-5 max-w-4xl">
-          <p className="prose mx-auto mt-3 text-center text-lg dark:prose-invert">
+          <p className="prose mx-auto text-center text-lg dark:prose-invert">
             This course is designed to be the absolute best starting point for Web Developers
             looking to learn Web3 Development. Solana is the ideal network for starting your Web3
             journey because of its high speed, low cost, energy efficiency, and more.
           </p>
         </div>
-
-        <Banner />
 
         <Table />
       </div>
